refactor: 탐색탭 내 유저 템플릿 조회 쿼리 별도 파일로 분리
